Complete refactor of CLI using Sade + expose page building api through src/build.js
