+++ examples/basic/package.json
@@ -4,8 +4,8 @@
     "description": "Basic example using Dhow as a Static Site Generator",
     "main": "index.js",
     "scripts": {
-        "dev": "dhow -w",
-        "build": "dhow"
+        "dev": "dhow dev",
+        "build": "dhow build"
     },
     "author": "",
     "license": "MIT",
