+++ examples/blog/package.json
@@ -4,8 +4,8 @@
     "description": "",
     "main": "index.js",
     "scripts": {
-        "dev": "dhow -w",
-        "build": "dhow"
+        "dev": "dhow dev",
+        "build": "dhow build"
     },
     "keywords": [],
     "author": "",
@@ -13,7 +13,7 @@
     "dependencies": {
         "dhow": "^2.0.0",
         "gray-matter": "^4.0.2",
-        "marked": "^1.1.0",
-        "prismjs": "^1.20.0"
+        "marked": "^1.1.1",
+        "prismjs": "^1.21.0"
     }
 }
