+++ examples/portfolio/public/fonts/Inter-Italic.woff2
